Moved variant tests to new format
